handle delete task

+++ src/tasks/tasks.service.ts
@@ -48,8 +48,8 @@ export class TasksService {
     return task;
   }
 
-  async deleteTask(id: number): Promise<object> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<object> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with id ${id} was not found`);
     } else {
